+++ pkg/lib/cisco/asa_test.go
@@ -33,7 +33,7 @@ func (b mockBase) SetSecret(secret string) {
 
 }
 
-func (b mockBase) SetTimeout(timeout int) {
+func (b mockBase) SetTimeout(timeout uint8) {
 }
 
 func TestASADevice_Connect(t *testing.T) {
